radiology-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
celery==5.3.6
redis==5.0.1
pillow==10.2.0
python-multipart==0.0.6
aiofiles==23.2.1
